backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.6
defusedxml==0.7.1
Django==4.2.23
django-cors-headers==4.7.0
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.16.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dotenv==0.9.9
flake8==7.3.0
gunicorn==23.0.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
psycopg2==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
